+++ .nycrc.yml
@@ -1,9 +1,9 @@
 exclude:
-  - '**/*.spec.ts'
+  - test
+  - dist
 reporter:
   - lcov
   - text
-branches: 100
-lines: 100
-functions: 100
-statements: 100
+all: true
+extends: '@istanbuljs/nyc-config-typescript'
+
