+++ package.json
@@ -12,7 +12,7 @@
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "docs": "shx rm -rf docs && typedoc --plugin typedoc-plugin-markdown --namedAnchors --hideBreadcrumbs",
     "lint": "eslint --ext .ts src/.",
-    "test": "shx rm -rf .nyc_output coverage && nyc mocha -r ts-node/register src/**/*.spec.ts",
+    "test": "shx rm -rf .nyc_output coverage && nyc mocha -r ts-node/register test/**/*.ts",
     "prepare": "yarn build",
     "prepublishOnly": "yarn lint && yarn test",
     "preversion": "yarn lint && yarn test && yarn build",
@@ -21,6 +21,7 @@
   },
   "dependencies": {},
   "devDependencies": {
+    "@istanbuljs/nyc-config-typescript": "^1.0.1",
     "@types/chai": "^4.2.11",
     "@types/expect": "^24.3.0",
     "@types/mocha": "^7.0.2",
@@ -36,6 +37,7 @@
     "mocha-lcov-reporter": "^1.3.0",
     "nyc": "^15.0.1",
     "shx": "^0.3.2",
+    "ts-node": "^8.8.2",
     "ts-node-dev": "^1.0.0-pre.44",
     "typedoc": "^0.17.4",
     "typedoc-plugin-markdown": "^2.2.17",
